refactor(subagent): extract constants + fork types from subagent.ts (#829)
subagent.ts carried three concerns: the default budget constants, the
fork-time option/parameter types, and the SubagentManager class itself.
Move the first two into the in-progress src/agent/subagent/ split,
following the convention already established there by handle.ts,
result.ts, pause-window.ts et al.

  src/agent/subagent/constants.ts   <- subagent.ts:57-193  (verbatim)
  src/agent/subagent/fork-types.ts  <- subagent.ts:195-415 (verbatim)

Both moved bodies are byte-identical to the originals, with exactly one
required edit: the inline dynamic-import type reference in
SubagentManagerOptions.onSubagentSucceeded drops its './subagent/' path
segment now that the declaration lives inside that directory.

All ten moved symbols are re-exported from subagent.ts unchanged, so the
33 existing `from './subagent.js'` importers are untouched. Five imports
in subagent.ts became unused once the moved code left and were dropped
(ZodType, IAgentSession, ElicitationRequest/Result, env, PhaseRole).

The SubagentManager class body is provably unmodified: all 749 lines
from `export class SubagentManager {` to EOF are byte-identical to the
pre-refactor file. forkSubagent's body and buildChildConfig are
deliberately out of scope (field-order hazards documented at :807-811
and :936 make that a separate change).

Behavior-preserving; no runtime change.
